Fix: add implementationVersionState to SafeInfo
