+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@gnosis.pm/safe-react-gateway-sdk",
-  "version": "3.2.1",
+  "version": "3.2.2",
   "main": "dist/index.min.js",
   "types": "dist/index.d.ts",
   "files": [
